[lua] Fix BattlefieldQuest Requirements
Adds in the RequiredVar / RequiredValue parameters to BattlefieldQuest.  Since this object inherits from the base Battlefield, it did not have direct access to the RequiredVar/Value params.  This meant that a player could endlessly enter into a Quest Related BCNM even after completion and no longer meeting the requirements specified in the lua for that BCNM.