[lua] Fix Transport Event in Selbina
Fixes a logical error in the transport event for Selbina <-> Mhaura boat.  Previously was returning early if the player was NOT in an event, but this is backwards - We only want to return early if the player already is in an event since the pirate boat and the main boat triggers fire at the same time.