[cpp] Fix BST Stay Command
Adjusts the checks in pet_controller.cpp during DoRoamTick() to properly check for BST Pets.  Previously the logic was only checking for TYPE_PET before moving into the pathfinding logic.  Charmed pets are not TYPE_PET though, and and as a result, if a charmed pet was told to Stay, it would ignore it and continue to follow the player.  This PR adds an additional check to account for charmed pets.